upload and displaying markdown from storage
