+++ src/routes/blog/test.svelte
@@ -3,8 +3,6 @@
     const posts = await fetch('/api/posts.json')
     const allPosts = await posts.json()
 
-    console.log('hello')
-    console.log(allPosts)
     return {
       props: {
         posts: allPosts,
@@ -18,22 +16,6 @@
   export let posts
   import SvelteMarkdown from 'svelte-markdown'
   const source = posts.markdown
-  const sourceTest = `
-  # This is a header
-
-This is a paragraph.
-
-* This is a list
-* With two items
-  1. And a sublist
-  2. That is ordered
-    * With another
-    * Sublist inside
-
-| And this is | A table |
-|-------------|---------|
-| With two    | columns |
-`
 </script>
 
 <article class="prose prose-pre:text-white">
