--- src/routes/blog/uses-copy.md
@@ -1,46 +1,0 @@
-# Uses Copy
-
-**Here's some stuff that sucks freaking ass**
-
-- SvelteKit
-- SvelteKit
-- SvelteKit
-- SvelteKit
-- SvelteKit
-- SvelteKit
-- SvelteKit
-- SvelteKit
-- SvelteKit
-- SvelteKit
-- SvelteKit
-- SvelteKit
-- SvelteKit
-- SvelteKit
-- VS Code
-- Emojis 😎
-- VS Code
-- Emojis 😎
-- VS Code
-- Emojis 😎
-- VS Code
-- Emojis 😎
-- VS Code
-- Emojis 😎
-- VS Code
-- Emojis 😎
-- VS Code
-- Emojis 😎
-- VS Code
-- Emojis 😎
-- VS Code
-- Emojis 😎
-- VS Code
-- Emojis 😎
-- VS Code
-- Emojis 😎
-- VS Code
-- Emojis 😎
-- VS Code
-- Emojis 😎
-- VS Code
-- Emojis 😎
